Adding ability to limit class properties with a prefix. (If you don't want NSObject and UIViewController properties)

+++ SigmaticOC/String/SOCStringExtension.m
@@ -127,7 +127,7 @@
 }
 
 - (void)forEachCharacter:(void (^)(NSString *))block {
-    unsigned int len = [self length];
+    NSUInteger len = [self length];
     unichar buffer[len];
     [self getCharacters:buffer range:NSMakeRange(0, len)];
